refactoring populator
