+++ pkg/lib/registry/registry_test.go
@@ -2,31 +2,20 @@ package registry
 
 import (
 	"context"
-	"database/sql"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"math/rand"
-	"os"
-	"path/filepath"
 	"testing"
 	"testing/fstest"
 	"time"
 
 	"github.com/stretchr/testify/require"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"sigs.k8s.io/yaml"
 
 	"github.com/operator-framework/operator-registry/alpha/model"
 	"github.com/operator-framework/operator-registry/alpha/property"
 	"github.com/operator-framework/operator-registry/pkg/image"
-	"github.com/operator-framework/operator-registry/pkg/lib/bundle"
 	"github.com/operator-framework/operator-registry/pkg/registry"
-	"github.com/operator-framework/operator-registry/pkg/registry/registryfakes"
-	"github.com/operator-framework/operator-registry/pkg/sqlite"
-	"github.com/operator-framework/operator-registry/pkg/sqlite/sqlitefakes"
 )
 
 func fakeBundlePathFromName(name string) string {
@@ -257,499 +246,4 @@ func TestUnpackImage(t *testing.T) {
 
 func init() {
 	rand.Seed(time.Now().UTC().UnixNano())
-}
-
-func CreateTestDb(t *testing.T) (*sql.DB, func()) {
-	dbName := fmt.Sprintf("test-%d.db", rand.Int())
-
-	db, err := sqlite.Open(dbName)
-	require.NoError(t, err)
-
-	return db, func() {
-		defer func() {
-			if err := os.Remove(dbName); err != nil {
-				t.Fatal(err)
-			}
-		}()
-		if err := db.Close(); err != nil {
-			t.Fatal(err)
-		}
-	}
-}
-
-func newUnpackedTestBundle(dir, name string, csvSpec json.RawMessage, annotations registry.Annotations, overwrite bool) (string, func(), error) {
-	bundleDir := filepath.Join(dir, fmt.Sprintf("%s-%s", annotations.PackageName, name))
-	cleanup := func() {
-		os.RemoveAll(bundleDir)
-	}
-
-	if overwrite {
-		os.RemoveAll(bundleDir)
-	}
-	if err := os.Mkdir(bundleDir, 0755); err != nil {
-		return bundleDir, cleanup, err
-	}
-	if err := os.Mkdir(filepath.Join(bundleDir, bundle.ManifestsDir), 0755); err != nil {
-		return bundleDir, cleanup, err
-	}
-	if err := os.Mkdir(filepath.Join(bundleDir, bundle.MetadataDir), 0755); err != nil {
-		return bundleDir, cleanup, err
-	}
-	if len(csvSpec) == 0 {
-		csvSpec = json.RawMessage(`{}`)
-	}
-
-	rawCSV, err := json.Marshal(registry.ClusterServiceVersion{
-		TypeMeta: v1.TypeMeta{
-			Kind: sqlite.ClusterServiceVersionKind,
-		},
-		ObjectMeta: v1.ObjectMeta{
-			Name: name,
-		},
-		Spec: csvSpec,
-	})
-	if err != nil {
-		return bundleDir, cleanup, err
-	}
-
-	rawObj := unstructured.Unstructured{}
-	if err := json.Unmarshal(rawCSV, &rawObj); err != nil {
-		return bundleDir, cleanup, err
-	}
-	rawObj.SetCreationTimestamp(v1.Time{})
-
-	jsonout, err := rawObj.MarshalJSON()
-	out, err := yaml.JSONToYAML(jsonout)
-	if err != nil {
-		return bundleDir, cleanup, err
-	}
-	if err := ioutil.WriteFile(filepath.Join(bundleDir, bundle.ManifestsDir, "csv.yaml"), out, 0666); err != nil {
-		return bundleDir, cleanup, err
-	}
-
-	out, err = yaml.Marshal(registry.AnnotationsFile{Annotations: annotations})
-	if err != nil {
-		return bundleDir, cleanup, err
-	}
-	if err := ioutil.WriteFile(filepath.Join(bundleDir, bundle.MetadataDir, "annotations.yaml"), out, 0666); err != nil {
-		return bundleDir, cleanup, err
-	}
-	return bundleDir, cleanup, nil
-}
-
-type bundleDir struct {
-	version     string
-	csvSpec     json.RawMessage
-	annotations registry.Annotations
-}
-
-func TestCheckForBundles(t *testing.T) {
-	type step struct {
-		bundles  map[string]bundleDir
-		action   int
-		expected map[string]*registry.Package
-	}
-	const (
-		actionAdd = iota
-		actionDeprecate
-		actionOverwrite
-	)
-	tests := []struct {
-		description string
-		steps       []step
-		wantErr     error
-		init        func() (*sql.DB, func())
-	}{
-		{
-			// 1.1.0 -> 1.0.0         pruned    channel 1
-			//        		\-> 1.2.0 ok        channel 2
-			description: "partialPruning",
-			steps: []step{
-				{
-					bundles: map[string]bundleDir{
-						"unorderedReplaces-1.1.0": {
-							csvSpec: json.RawMessage(`{"version":"1.1.0"}`),
-							annotations: registry.Annotations{
-								PackageName:        "testpkg",
-								Channels:           "stable,alpha",
-								DefaultChannelName: "stable",
-							},
-							version: "1.1.0",
-						},
-						"unorderedReplaces-1.0.0": {
-							csvSpec: json.RawMessage(`{"version":"1.0.0","replaces":"unorderedReplaces-1.1.0"}`),
-							annotations: registry.Annotations{
-								PackageName:        "testpkg",
-								Channels:           "stable,alpha",
-								DefaultChannelName: "stable",
-							},
-							version: "1.0.0",
-						},
-						"unorderedReplaces-1.2.0": {
-							csvSpec: json.RawMessage(`{"version":"1.2.0","replaces":"unorderedReplaces-1.0.0"}`),
-							annotations: registry.Annotations{
-								PackageName:        "testpkg",
-								Channels:           "alpha",
-								DefaultChannelName: "stable",
-							},
-							version: "1.2.0",
-						},
-					},
-					action: actionAdd,
-					expected: map[string]*registry.Package{
-						"testpkg": {
-							Name: "testpkg",
-							Channels: map[string]registry.Channel{
-								"alpha": {
-									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-										registry.BundleKey{
-											BundlePath: "unorderedReplaces-1.0.0",
-											Version:    "1.0.0",
-											CsvName:    "unorderedReplaces-1.0.0",
-										}: nil,
-										registry.BundleKey{
-											BundlePath: "unorderedReplaces-1.1.0",
-											Version:    "1.1.0",
-											CsvName:    "unorderedReplaces-1.1.0",
-										}: nil,
-										registry.BundleKey{
-											BundlePath: "unorderedReplaces-1.2.0",
-											Version:    "1.2.0",
-											CsvName:    "unorderedReplaces-1.2.0",
-										}: nil,
-									},
-								},
-								"stable": {
-									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-										registry.BundleKey{
-											BundlePath: "unorderedReplaces-1.0.0",
-											Version:    "1.0.0",
-											CsvName:    "unorderedReplaces-1.0.0",
-										}: nil,
-										registry.BundleKey{
-											BundlePath: "unorderedReplaces-1.1.0",
-											Version:    "1.1.0",
-											CsvName:    "unorderedReplaces-1.1.0",
-										}: nil,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: fmt.Errorf("added bundle unorderedReplaces-1.0.0 pruned from package testpkg, channel stable: this may be due to incorrect channel head (unorderedReplaces-1.1.0)"),
-		},
-		{
-			description: "ignoreDeprecated",
-			steps: []step{
-				{
-					bundles: map[string]bundleDir{
-						"ignoreDeprecated-1.0.0": {
-							csvSpec: json.RawMessage(`{"version":"1.0.0"}`),
-							annotations: registry.Annotations{
-								PackageName: "testpkg",
-								Channels:    "stable",
-							},
-							version: "1.0.0",
-						},
-						"ignoreDeprecated-1.1.0": {
-							csvSpec: json.RawMessage(`{"version":"1.1.0","replaces":"ignoreDeprecated-1.0.0"}`),
-							annotations: registry.Annotations{
-								PackageName: "testpkg",
-								Channels:    "stable",
-							},
-							version: "1.1.0",
-						},
-						"ignoreDeprecated-1.2.0": {
-							csvSpec: json.RawMessage(`{"version":"1.2.0","replaces":"ignoreDeprecated-1.1.0"}`),
-							annotations: registry.Annotations{
-								PackageName: "testpkg",
-								Channels:    "stable",
-							},
-							version: "1.2.0",
-						},
-					},
-					action: actionAdd,
-					expected: map[string]*registry.Package{
-						"testpkg": {
-							Name: "testpkg",
-							Channels: map[string]registry.Channel{
-								"stable": {
-									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-										registry.BundleKey{
-											BundlePath: "ignoreDeprecated-1.0.0",
-											Version:    "1.0.0",
-											CsvName:    "ignoreDeprecated-1.0.0",
-										}: nil,
-										registry.BundleKey{
-											BundlePath: "ignoreDeprecated-1.1.0",
-											Version:    "1.1.0",
-											CsvName:    "ignoreDeprecated-1.1.0",
-										}: nil,
-										registry.BundleKey{
-											BundlePath: "ignoreDeprecated-1.2.0",
-											Version:    "1.2.0",
-											CsvName:    "ignoreDeprecated-1.2.0",
-										}: nil,
-									},
-								},
-							},
-						},
-					},
-				},
-				{
-					bundles: map[string]bundleDir{
-						"ignoreDeprecated-1.1.0": {},
-					},
-					action: actionDeprecate,
-					expected: map[string]*registry.Package{
-						"testpkg": {
-							Name: "testpkg",
-							Channels: map[string]registry.Channel{
-								"stable": {
-									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-										registry.BundleKey{
-											BundlePath: "ignoreDeprecated-1.1.0",
-											Version:    "1.1.0",
-											CsvName:    "ignoreDeprecated-1.1.0",
-										}: nil,
-										registry.BundleKey{
-											BundlePath: "ignoreDeprecated-1.2.0",
-											Version:    "1.2.0",
-											CsvName:    "ignoreDeprecated-1.2.0",
-										}: nil,
-									},
-								},
-							},
-						},
-					},
-				},
-				{
-					bundles: map[string]bundleDir{
-						"ignoreDeprecated-1.2.0": {
-							csvSpec: json.RawMessage(`{"version":"1.2.0","replaces":""}`),
-							annotations: registry.Annotations{
-								PackageName:        "testpkg",
-								Channels:           "alpha",
-								DefaultChannelName: "alpha",
-							},
-							version: "1.2.0",
-						},
-					},
-					action: actionOverwrite,
-					expected: map[string]*registry.Package{
-						"testpkg": {
-							Name: "testpkg",
-							Channels: map[string]registry.Channel{
-								"stable": {
-									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-										registry.BundleKey{
-											BundlePath: "ignoreDeprecated-1.1.0",
-											Version:    "1.1.0",
-											CsvName:    "ignoreDeprecated-1.1.0",
-										}: nil,
-									},
-								},
-								"alpha": {
-									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-										registry.BundleKey{
-											BundlePath: "ignoreDeprecated-1.2.0-overwrite",
-											Version:    "1.2.0",
-											CsvName:    "ignoreDeprecated-1.2.0",
-										}: nil,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.description, func(t *testing.T) {
-			tmpdir, err := os.MkdirTemp(".", "tmpdir-*")
-			defer os.RemoveAll(tmpdir)
-			db, cleanup := CreateTestDb(t)
-			defer cleanup()
-			load, err := sqlite.NewSQLLiteLoader(db)
-			require.NoError(t, err)
-			require.NoError(t, load.Migrate(context.TODO()))
-			query := sqlite.NewSQLLiteQuerierFromDb(db)
-			graphLoader, err := sqlite.NewSQLGraphLoaderFromDB(db)
-			require.NoError(t, err)
-
-			for _, step := range tt.steps {
-				switch step.action {
-				case actionDeprecate:
-					for deprecate := range step.bundles {
-						require.NoError(t, load.DeprecateBundle(deprecate))
-					}
-				case actionAdd, actionOverwrite:
-					overwriteRefs := map[string][]string{}
-					refs := map[image.Reference]string{}
-					for name, b := range step.bundles {
-						dir, _, err := newUnpackedTestBundle(tmpdir, name, b.csvSpec, b.annotations, true)
-						require.NoError(t, err)
-
-						// refs to be added
-						bundleImage := name
-
-						// bundles to remove for overwrite. Only one per package is permitted.
-						if step.action == actionOverwrite {
-							bundleImage += "-overwrite"
-							img, err := registry.NewImageInput(image.SimpleReference(bundleImage), dir)
-							require.NoError(t, err)
-							overwriteRefs[img.Bundle.Package] = append(overwriteRefs[img.Bundle.Package], name)
-						}
-						refs[image.SimpleReference(bundleImage)] = dir
-
-					}
-					require.NoError(t, registry.NewDirectoryPopulator(
-						load,
-						graphLoader,
-						query,
-						refs,
-						overwriteRefs).Populate(registry.ReplacesMode))
-
-				}
-				err = checkForBundles(context.TODO(), query, graphLoader, step.expected)
-				if tt.wantErr == nil {
-					require.NoError(t, err, fmt.Sprintf("%d", step.action))
-					continue
-				}
-				require.EqualError(t, err, tt.wantErr.Error())
-			}
-		})
-	}
-}
-
-func TestDeprecated(t *testing.T) {
-	deprecated := map[string]bool{
-		"deprecatedBundle": true,
-		"otherBundle":      false,
-	}
-	q := &sqlitefakes.FakeQuerier{
-		QueryContextStub: func(ctx context.Context, query string, args ...interface{}) (sqlite.RowScanner, error) {
-			bundleName := args[2].(string)
-			if len(bundleName) == 0 {
-				return nil, fmt.Errorf("empty bundle name")
-			}
-			hasNext := true
-			return &sqlitefakes.FakeRowScanner{ScanStub: func(args ...interface{}) error {
-				if deprecated[bundleName] {
-					*args[0].(*sql.NullString) = sql.NullString{
-						String: registry.DeprecatedType,
-						Valid:  true,
-					}
-					*args[1].(*sql.NullString) = sql.NullString{
-						Valid: true,
-					}
-				}
-				return nil
-			},
-				NextStub: func() bool {
-					if hasNext {
-						hasNext = false
-						return true
-					}
-					return false
-				},
-			}, nil
-		},
-	}
-
-	querier := sqlite.NewSQLLiteQuerierFromDBQuerier(q)
-
-	_, err := isDeprecated(context.TODO(), querier, registry.BundleKey{})
-	require.Error(t, err)
-
-	for b := range deprecated {
-		isDeprecated, err := isDeprecated(context.TODO(), querier, registry.BundleKey{BundlePath: b})
-		require.NoError(t, err)
-		require.Equal(t, deprecated[b], isDeprecated)
-	}
-}
-
-func TestExpectedGraphBundles(t *testing.T) {
-	testBundle, err := registry.NewBundleFromStrings("testBundle", "0.0.1", "testPkg", "default", "default", "")
-	require.NoError(t, err)
-	testBundle.BundleImage = "testImage"
-	testBundleKey := registry.BundleKey{
-		BundlePath: testBundle.BundleImage,
-		Version:    "0.0.1",
-		CsvName:    testBundle.Name,
-	}
-	newTestPackage := func(name string, channelEntries map[string]registry.BundleKey) *registry.Package {
-		channels := map[string]registry.Channel{}
-		for channelName, node := range channelEntries {
-			if _, ok := channels[channelName]; !ok {
-				channels[channelName] = registry.Channel{
-					Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{},
-				}
-			}
-			channels[channelName].Nodes[node] = nil
-		}
-		return &registry.Package{
-			Name:     name,
-			Channels: channels,
-		}
-	}
-
-	tests := []struct {
-		description      string
-		graphLoader      registry.GraphLoader
-		bundles          []*registry.Bundle
-		overwrite        bool
-		wantErr          error
-		wantGraphBundles map[string]*registry.Package
-	}{
-		{
-			description: "GraphLoaderError",
-			graphLoader: &registryfakes.FakeGraphLoader{GenerateStub: func(string) (*registry.Package, error) { return nil, fmt.Errorf("graphLoader error") }},
-			bundles:     []*registry.Bundle{testBundle},
-			wantErr:     fmt.Errorf("graphLoader error"),
-		},
-		{
-			description: "NewPackage",
-			graphLoader: &registryfakes.FakeGraphLoader{GenerateStub: func(string) (*registry.Package, error) { return nil, registry.ErrPackageNotInDatabase }},
-			bundles:     []*registry.Bundle{testBundle},
-			wantGraphBundles: map[string]*registry.Package{
-				"testPkg": newTestPackage("testPkg", map[string]registry.BundleKey{"default": testBundleKey}),
-			},
-		},
-		{
-			description: "OverwriteWithoutFlag",
-			graphLoader: &registryfakes.FakeGraphLoader{GenerateStub: func(string) (*registry.Package, error) {
-				return newTestPackage("testPkg", map[string]registry.BundleKey{"alpha": testBundleKey}), nil
-			}},
-			bundles: []*registry.Bundle{testBundle},
-			wantErr: registry.BundleImageAlreadyAddedErr{ErrorString: fmt.Sprintf("Bundle %s already exists", testBundle.BundleImage)},
-		},
-		{
-			description: "OverwriteWithFlag",
-			graphLoader: &registryfakes.FakeGraphLoader{GenerateStub: func(string) (*registry.Package, error) {
-				return newTestPackage("testPkg", map[string]registry.BundleKey{"alpha": testBundleKey}), nil
-			}},
-			bundles:   []*registry.Bundle{testBundle},
-			overwrite: true,
-			wantGraphBundles: map[string]*registry.Package{
-				"testPkg": newTestPackage("testPkg", map[string]registry.BundleKey{"default": testBundleKey}),
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.description, func(t *testing.T) {
-			graphBundles, err := expectedGraphBundles(tt.bundles, tt.graphLoader, tt.overwrite)
-			if tt.wantErr != nil {
-				require.EqualError(t, err, tt.wantErr.Error())
-				return
-			}
-			require.NoError(t, err)
-
-			require.EqualValues(t, graphBundles, tt.wantGraphBundles)
-		})
-	}
 }
