+++ pkg/sqlite/remove_test.go
@@ -34,7 +34,7 @@ func TestRemover(t *testing.T) {
 			map[image.Reference]string{
 				image.SimpleReference("quay.io/test/" + name): "../../bundles/" + name,
 			},
-			nil).Populate(registry.ReplacesMode)
+			nil, registry.ReplacesMode).Populate(context.TODO())
 	}
 	for _, name := range []string{"etcd.0.9.0", "etcd.0.9.2", "prometheus.0.14.0", "prometheus.0.15.0", "prometheus.0.22.2"} {
 		require.NoError(t, populate(name))
